requirements.txt
absl-py==1.4.0
array-record==0.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
backcall==0.2.0
bokeh==3.1.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudml-hypertune==0.1.0.dev6
cloudpickle==3.0.0
contourpy==1.1.1
crepe==0.0.12
cycler==0.12.1
ddsp==3.7.0
decorator==5.1.1
dill==0.3.4
dm-tree==0.1.8
etils==1.3.0
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.1
future==1.0.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.19.1
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.1
grpcio==1.66.0
h5py==3.11.0
hmmlearn==0.2.7
idna==3.7
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
intervaltree==3.1.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
keras==2.11.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.0
llvmlite==0.41.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mido==1.2.10
mir_eval==0.7
msgpack==1.0.8
note-seq==0.0.3
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
pretty_midi==0.2.10
promise==2.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.19.6
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.5
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.2.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.13.0
tensorflow-probability==0.19.0
tensorflowjs==3.18.0
termcolor==2.4.0
tflite-support==0.1.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
xyzservices==2024.6.0
zipp==3.20.0
